backend/auction_images/serializers.py
# authentication/serializers.py
from rest_framework import serializers
from .models import *

class ReactSerializer(serializers.ModelSerializer):
    class Meta:
        model = React
        fields = ['post_id', 'image']

