backend/requirements.txt
anyio==4.2.0
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Django==4.2.5
django-cors-headers==4.3.0
django-crispy-forms==2.1
djangorestframework==3.14.0
mysqlclient==2.2.0
Pillow==10.0.1


